test(storage): split unit tests for emulator
This is just refactoring / cleanup to make the next round of changes
easier to isolate.
